# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Squawk social media API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  website?: string | null;
  isVerified?: boolean;
  followersCount: number;
  followingCount: number;
  postsCount: number;
  isFollowing: boolean;
  createdAt: string;
}

export interface UserSummary {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  isVerified?: boolean;
  isFollowing: boolean;
  /** @nullable */
  bio?: string | null;
}

export interface OnboardInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName: string;
  bio?: string;
  avatarUrl?: string;
}

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface FollowResult {
  isFollowing: boolean;
  followersCount: number;
}

export type PostMediaType = typeof PostMediaType[keyof typeof PostMediaType];


export const PostMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface Post {
  id: number;
  authorId: number;
  author: UserSummary;
  /** @nullable */
  caption?: string | null;
  mediaUrl: string;
  mediaType: PostMediaType;
  /** @nullable */
  thumbnailUrl?: string | null;
  hashtags?: string[];
  likesCount: number;
  commentsCount: number;
  viewsCount: number;
  isLiked: boolean;
  isSaved: boolean;
  createdAt: string;
}

export type PostInputMediaType = typeof PostInputMediaType[keyof typeof PostInputMediaType];


export const PostInputMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface PostInput {
  /** @nullable */
  caption?: string | null;
  mediaUrl: string;
  mediaType: PostInputMediaType;
  /** @nullable */
  thumbnailUrl?: string | null;
  hashtags?: string[];
}

export interface PostPage {
  posts: Post[];
  hasMore: boolean;
  /** @nullable */
  nextCursor?: number | null;
}

export interface LikeResult {
  isLiked: boolean;
  likesCount: number;
}

export interface SaveResult {
  isSaved: boolean;
}

export interface Comment {
  id: number;
  postId: number;
  authorId: number;
  author: UserSummary;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
}

export type StoryMediaType = typeof StoryMediaType[keyof typeof StoryMediaType];


export const StoryMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface Story {
  id: number;
  authorId: number;
  author: UserSummary;
  mediaUrl: string;
  mediaType: StoryMediaType;
  viewsCount: number;
  createdAt: string;
  expiresAt: string;
  isViewed: boolean;
}

export interface StoryGroup {
  user: UserSummary;
  stories: Story[];
  hasUnviewed: boolean;
}

export type StoryInputMediaType = typeof StoryInputMediaType[keyof typeof StoryInputMediaType];


export const StoryInputMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface StoryInput {
  mediaUrl: string;
  mediaType: StoryInputMediaType;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface FeedStats {
  totalLikes: number;
  totalPosts: number;
  totalFollowers: number;
  totalViews: number;
  recentActivity: ActivityItem[];
}

export interface SearchResults {
  users: UserSummary[];
  posts: Post[];
}

export interface HashtagSummary {
  tag: string;
  postCount: number;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  sender: UserSummary;
  content: string;
  /** @nullable */
  mediaUrl?: string | null;
  createdAt: string;
}

export interface Conversation {
  id: number;
  participants: UserSummary[];
  lastMessage?: Message;
  unreadCount?: number;
  createdAt: string;
}

export interface ConversationInput {
  recipientUsername: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  mediaUrl?: string | null;
}

export interface MessagePage {
  messages: Message[];
  hasMore: boolean;
  /** @nullable */
  nextCursor?: number | null;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  like: 'like',
  comment: 'comment',
  follow: 'follow',
  mention: 'mention',
  message: 'message',
} as const;

export interface Notification {
  id: number;
  type: NotificationType;
  isRead: boolean;
  actor?: UserSummary;
  /** @nullable */
  postId?: number | null;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type GetSuggestedUsersParams = {
limit?: number;
};

export type ListPostsParams = {
/**
 * @nullable
 */
cursor?: number | null;
limit?: number;
};

export type GetTrendingPostsParams = {
limit?: number;
};

export type GetFeedParams = {
/**
 * @nullable
 */
cursor?: number | null;
limit?: number;
};

export type SearchParams = {
q: string;
type?: SearchType;
};

export type SearchType = typeof SearchType[keyof typeof SearchType];


export const SearchType = {
  users: 'users',
  posts: 'posts',
  all: 'all',
} as const;

export type GetNotificationsParams = {
unreadOnly?: boolean;
limit?: number;
};