# .migration-backup/lib/api-zod/src/generated/types/postPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Squawk social media API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface PostPage {
  posts: Post[];
  hasMore: boolean;
  /** @nullable */
  nextCursor?: number | null;
}